+++ CHANGELOG.md
@@ -11,7 +11,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add more fields to `dns`, `conn`, `http`
 - Add common fields to network events.
-- Add Packet store
+- Add Packet store.
+- Add GraphQL API for Packet.
 
 ### Changed